feat: Added add to & remove from basket functionality to product cards
Users can now add or remove a product to/from their basket from the product card. The total number of items in the basket is displated in the basket header component. Future improvements to display the total price in header component and to build basket flyout component to view contents all together.