feat: Updated basket for rucher functionality
Users can now add and remove individual quantities of an item to their basket or delete a whole quantity. The basket dropdown is scrollable and a dark overlay covers the main products page when the basket is open. The basket no longer shows an 'empty' message if clicked while not items in it.